// src/generated/models/cr2b6_operationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface cr2b6_operations {
  cr2b6_batch?: string;
  cr2b6_batchname?: string;
  cr2b6_description?: string;
  cr2b6_endtime: Date;
  cr2b6_id: string;
  cr2b6_operationid: string;
  cr2b6_starttime: Date;
  cr2b6_system?: string;
  cr2b6_systemname?: string;
  cr2b6_type?: string;
  cr2b6_typename?: string;
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: Date;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: number;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: Date;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: Date;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: string;
  statecodename?: string;
  statuscode?: string;
  statuscodename?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
}